First part of the suggestion implementation in PR 649
